Correction and suggestions for: 10.17504/protocols.io.bf4vjqw6;
Samples under: 10.17504/protocols.io.bf4vjqw6; The specifed lipidomics protocol does not have a liquid chromatography step and instead uses flow injection analysis and mass spec (FIA-MS) as specified in the protocol. This is my suggest dataset_type as that was specified by the data provider. Technically the "LC-30AD" lc instrument, is not an an LC instrument but a liquid solvent delivery pump. Leaving as is right now while we can strategy if we need new fields.